core/requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
Django==4.0.6
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
inflection==0.3.1
iniconfig==1.1.1
lazy-object-proxy==1.7.1
model-bakery==1.6.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
wrapt==1.14.1
